Move files into movegen directory
