+++ chess_board/pinmask.hpp
