+++ chess_board/slider_attacks/kindergarten.hpp
