+++ main.cpp
@@ -1,5 +1,7 @@
 #include <iostream>
 
+#include "chess_board/board.hpp"
+
 int main () {
     std::cout << "id name Motor " << std::endl;
     std::cout << "id author Martin Novak" << std::endl;
